backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
pgvector==0.2.4

# LLM & Embeddings
openai==1.12.0
google-generativeai==0.3.2
langchain==0.1.5
langchain-openai==0.0.5
tiktoken==0.5.2

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0

# Development
ruff==0.1.14
pytest==7.4.4
pytest-asyncio==0.23.3
